Fix not finding `rubocop/rake_task` in production
Heroku does not install gems from development/test groups.  We have
Rubocop defined only in the test group in the Gemfile.  Rake was still
trying to load the `ci` namespace even in production.

This modifies the `ci` namespace to only be defined in non-production
environments, so `rubocop/rake_task` will never not be found.